query builder: more cleanups
