+++ packages/legend-studio-plugin-query-builder/src/stores/QueryBuilderSetupState.ts
@@ -165,7 +165,7 @@ export class QueryBuilderSetupState {
   ) {
     this.setMapping(mapping);
     this.setRuntime(runtime);
-    this.queryBuilderState.initializeStateWithRawLambda(func);
+    this.queryBuilderState.init(func);
     this.setOnSaveQuery(onSave);
     yield this.queryBuilderState.setOpenQueryBuilder(true, {
       disableCompile: disableCompile,
